com.vmturbo.prometheus.config.manager/docker/requirements.txt
requests==2.24.0
pyyaml==5.3.1
python-consul2==0.1.4
